Move custom hooks for bags into the service facade, since it's all API related
